requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
Werkzeug==3.0.1
marshmallow==3.20.1
python-dateutil==2.8.2